requirements.txt
streamlit==1.37.1
google-genai>=0.2.0
pandas>=2.1.0
numpy>=1.25.0
